Create lib.rs and restructure crate
This (quite huge) commit makes present usable as a library. Doing this
required a lot of restructuring of the internal API, but the CLI remains
the same.

Along the way, I was able to remove quite a lot of unnecessary
allocations (clones and collects), which means the resulting binary
should be a bit faster - though I have not benchmarked it.